web/codebackup/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
Flask_Limiter==1.4
pycryptodome==3.9.8
PyYAML==5.3.1
